requirements.txt
youtube-transcript-api
langchain-community
python-dotenv
langchain
qdrant-client
fastembed